// frontend/src/src/lib/types.ts
export type StatusEmbarcacao = "DISPONIVEL" | "RESERVADA" | "EM_AULA" | "MANUTENCAO" | "INTERDITADA";
export type DiaSemana = "SEGUNDA" | "TERCA" | "QUARTA" | "QUINTA" | "SEXTA" | "SABADO" | "DOMINGO";
export type StatusReserva = "CONFIRMADA" | "CANCELADA" | "CONCLUIDA";
export type StatusOcorrencia = "ABERTA" | "EM_ANALISE" | "RESOLVIDA";
export type GravidadeOcorrencia = "BAIXA" | "MEDIA" | "ALTA";
export type StatusSlot = "EM_ANDAMENTO" | "PROXIMA" | "ENCERRADA" | "AGENDADA";